Corrections on activity 2
Corrected a 1 that should have been a -1 on the calculation of b x c at the beginning of the activity.
Corrected notation and process errors located by Choy on the reciprocal vector of C, having used a wrong vector as B therefore having a wrong result overall.
Also corrected spacing overall of the document source code.